Update school_subgroups_V.sql
I don't understand after line 30 till line 98, is this pulling from a nested table you are creating? If so, what is the "SPED" variable does that need to be included or is it part of SPPF..SPECIAL EDUCATION and not needed to be called out in the second select statement?
        ___These same questions apply to the gifted section.

What is the "p" for on line 98?

On lines 253-260 two questions: 1) I think the sum and isnull might be in the wrong order? 2) do you need to cast these as FLOAT to avoid problems?